LayoutTests/fast/dom/AbortSignal-reason-crash-expected.txt

Tests accessing the reason of a newly created AbortSignal.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new AbortController().signal.reason is undefined.
PASS successfullyParsed is true

TEST COMPLETE
